# Optimize ManifestResourceStream.Read(Span<byte>)